Add default user for agnoster theme to zshrc template.
Without a default user set, the LPROMPT looks like

  user@hostname > dir >

With the user set to the currently logged in user, the prompt becomes

  dir >

This can be useful when ssh-ing into remote machines, as the hostname
will appear. However, on a local machine the user most likely know who
they are and the name of their computer, so it is not required.