Allow for sudo with aliases.
Uses a custom function called 'do_sudo' in order to allow sudo
to be used with aliases. The function was acquired from here:
http://www.zsh.org/mla/users/2008/msg01229.html

This function is used instead of "alias sudo='sudo '" because the
alias solution does not work when the next alias contains 'nocorrect'.
Certain commands have these precommand modifiers by default in this
zsh configuration, such as mv and mkdir. As such, the alias solution
will cause these functions to be unusable, as zsh will try to run 'sudo
nocorrect mv ...', which responds with something similar to "sudo:
nocorrect command not found."