Alter agnoster-alt theme to differentiate ssh from non-default user.
If the current user of a prompt is not the default user, then the
username and hostname are displayed, with the background in black and
the text in default (white). However, if the user has ssh'd into the
machine, then the text has a cyan colour. The cyan colour takes
precedence over the default user switch.